Fix more mouse button options in outliner
The function "select_material" was getting the "id"s of the
faces and using them as "id"s of the other elements (for the
edge/vertex selection mode).

NOTE: Multiple mouse buttons options in the context menu of Outliner window.
The option Select by material in the Outliner window doesn't
select the right elements if the selection mode isn't Face or
Body. It was fixed and added new parameters
that enable us to add support for "add to selection" and "remove from
selection" in context menu.
(Thanks to deerwood that reported the bug
and suggested the new options) [Micheus]