MyController.m: Eliminate warnings
The first argument for NSLog() should be a format string, not an
arbitrary string.

The Gestalt() function is deprecated starting in OS X 10.8.
Since the only purpose of the call was to avoid running on too old
versions of OS X, we can simply remove the call since the deployment
version set in the binary will server the same purpose.