Translate update wpc_ncube_jp.lang (fixed)
wpc_ncube->cube_str->1 : Cube
wpc_ncube->make_ngon->2 : N-gon

In particular, I was returned to the original Cube string. It's because they are also used for menu item name or the object name. (If you open the English version the data that you have saved in other languages ​​if there is intended to prevent entry name of the geometry graph from being garbled as black squares)